[WIN32K] - Fix a memory leak in EngUnloadImage - add an ASSERT
svn path=/trunk/; revision=55970

+++ reactos/subsystems/win32/win32k/eng/pdevobj.c
@@ -190,7 +190,7 @@ PDEVOBJ_pdmMatchDevMode(
 {
     PGRAPHICS_DEVICE pGraphicsDevice;
     PDEVMODEW pdmCurrent;
-    INT i;
+    ULONG i;
     DWORD dwFields;
 
     pGraphicsDevice = ppdev->pGraphicsDevice;
